requirements.txt
agate==1.9.1
annotated-types==0.7.0
attrs==24.2.0
babel==2.16.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
daff==1.3.46
dbt-adapters==1.6.0
dbt-common==1.8.0
dbt-core==1.8.6
dbt-extractor==0.5.1
dbt-postgres==1.8.2
dbt-semantic-interfaces==0.5.1
deepdiff==7.0.1
idna==3.10
importlib-metadata==6.11.0
isodate==0.6.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
leather==0.4.0
Logbook==1.5.3
MarkupSafe==2.1.5
mashumaro==3.13.1
minimal-snowplow-tracker==0.0.2
more-itertools==10.5.0
msgpack==1.1.0
networkx==3.3
ordered-set==4.1.0
packaging==24.1
parsedatetime==2.6
pathspec==0.12.1
protobuf==4.25.4
psycopg2-binary==2.9.9
pydantic==2.9.1
pydantic_core==2.23.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
six==1.16.0
sqlparse==0.5.1
text-unidecode==1.3
typing_extensions==4.12.2
urllib3==2.2.3
zipp==3.20.2
agate==1.9.1
annotated-types==0.7.0
attrs==24.2.0
babel==2.16.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
daff==1.3.46
dbt-adapters==1.6.0
dbt-common==1.8.0
dbt-core==1.8.6
dbt-extractor==0.5.1
dbt-postgres==1.8.2
dbt-semantic-interfaces==0.5.1
deepdiff==7.0.1
idna==3.10
importlib-metadata==6.11.0
isodate==0.6.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
leather==0.4.0
Logbook==1.5.3
MarkupSafe==2.1.5
mashumaro==3.13.1
minimal-snowplow-tracker==0.0.2
more-itertools==10.5.0
msgpack==1.1.0
networkx==3.3
ordered-set==4.1.0
packaging==24.1
parsedatetime==2.6
pathspec==0.12.1
protobuf==4.25.4
psycopg2-binary==2.9.9
pydantic==2.9.1
pydantic_core==2.23.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
six==1.16.0
sqlparse==0.5.1
text-unidecode==1.3
typing_extensions==4.12.2
urllib3==2.2.3
zipp==3.20.2
